+++ lib/src/material_extensions/material_extensions.dart
@@ -1,14 +1,15 @@
 import 'package:coder_matthews_extensions/src/material_extensions/helper_classes.dart';
 import 'package:flutter/material.dart';
+export 'helper_classes.dart';
 
 extension GlobalKeyExn on GlobalKey {
-  ///Gets the [RelativeRect] and [RenderBox] in a nullable [PositionData] object on the widget attached to a given [GlobalKey]
-  PositionData? getKeyPosition(BuildContext context) {
+  ///Gets the [RelativeRect] and [RenderBox] in a nullable [WidgetPositionData] object on the widget attached to a given [GlobalKey]
+  WidgetPositionData? getKeyPosition(BuildContext context) {
     var button = currentContext?.findRenderObject() as RenderBox?;
     if (button == null) return null;
     var overlay = Overlay.maybeOf(context, rootOverlay: true)?.context.findRenderObject() as RenderBox?;
     if (overlay == null) return null;
-    return PositionData(
+    return WidgetPositionData(
       RelativeRect.fromRect(
         Rect.fromPoints(
           button.localToGlobal(button.size.topLeft(Offset.zero), ancestor: overlay),
@@ -19,4 +20,9 @@ extension GlobalKeyExn on GlobalKey {
       button,
     );
   }
+}
+
+extension BrightnessExt on Brightness {
+  bool get isLight => this == Brightness.light;
+  bool get isDark => this == Brightness.dark;
 }